Stripe API, ShipEngine API, AWS SES, Gson config implementation, Email Service, Account, Admin, Donate, Mock room, Art Process, Collections, Order entity, ModerationStatus, PaymentStatus enum

+++ src/main/java/com/example/artvswar/repository/specification/painting/PaintingPriceBetweenSpecification.java
@@ -14,8 +14,8 @@ public class PaintingPriceBetweenSpecification implements SpecificationProvider<
     public Specification<Painting> getSpecification(String[] fromToPrice) {
         return ((root, query, criteriaBuilder) ->
                 criteriaBuilder.between(root.get(FIELD_NAME),
-                        BigDecimal.valueOf(Integer.parseInt(fromToPrice[0])),
-                        BigDecimal.valueOf(Integer.parseInt(fromToPrice[1]))));
+                        BigDecimal.valueOf(Long.parseLong(fromToPrice[0])),
+                        BigDecimal.valueOf(Long.parseLong(fromToPrice[1]))));
     }
 
     @Override
